refactor(InputView): Input 관련 클래스 및 패키지 합병
